+++ src/index.rs
@@ -1,21 +1,25 @@
 use {
   self::{
     entry::{
-      BlockHashValue, Entry, InscriptionEntry, InscriptionEntryValue, InscriptionIdValue,
-      OutPointValue, SatPointValue, SatRange,
+      outpoint_prefix_end, BlockHashValue, Entry, InscriptionEntry, InscriptionEntryValue,
+      InscriptionIdValue, OutPointPrefix, OutPointPrefixValue, OutPointValue, SatPointValue,
+      SatRange,
     },
     updater::Updater,
   },
   super::*,
   crate::wallet::Wallet,
   bitcoin::block::Header,
-  bitcoincore_rpc::{json::GetBlockHeaderResult, Client},
+  bitcoincore_rpc::{
+    json::{GetBlockHeaderResult, GetRawTransactionResult},
+    Client,
+  },
   chrono::SubsecRound,
   indicatif::{ProgressBar, ProgressStyle},
   log::log_enabled,
   redb::{
-    Database, MultimapTable, MultimapTableDefinition, ReadableMultimapTable, ReadableTable, Table,
-    TableDefinition, WriteTransaction,
+    CompactionError, Database, MultimapTable, MultimapTableDefinition, ReadableMultimapTable,
+    ReadableTable, Table, TableDefinition, WriteTransaction,
   },
   std::collections::HashMap,
   std::io::{BufWriter, Read, Write},
@@ -48,6 +52,7 @@ define_table! { INSCRIPTION_ID_TO_INSCRIPTION_ENTRY, &InscriptionIdValue, Inscri
 define_table! { INSCRIPTION_ID_TO_SATPOINT, &InscriptionIdValue, &SatPointValue }
 define_table! { INSCRIPTION_NUMBER_TO_INSCRIPTION_ID, i64, &InscriptionIdValue }
 define_table! { OUTPOINT_TO_SAT_RANGES, &OutPointValue, &[u8] }
+define_table! { SAT_TO_OUTPOINT, u64, &OutPointPrefixValue }
 define_table! { OUTPOINT_TO_VALUE, &OutPointValue, u64}
 define_table! { REINSCRIPTION_ID_TO_SEQUENCE_NUMBER, &InscriptionIdValue, u64 }
 define_multimap_table! { SATPOINT_TO_INSCRIPTION_ID, &SatPointValue, &InscriptionIdValue }
@@ -243,9 +248,13 @@ impl Index {
         tx.open_table(STATISTIC_TO_COUNT)?
           .insert(&Statistic::Schema.key(), &SCHEMA_VERSION)?;
 
-        if options.index_sats {
+        if options.index_sats || options.index_utxos {
           tx.open_table(OUTPOINT_TO_SAT_RANGES)?
             .insert(&OutPoint::null().store(), [].as_slice())?;
+          if options.index_utxos {
+            tx.open_table(SAT_TO_OUTPOINT)?
+              .insert(0, &OutPoint::null().store().store())?;
+          }
         }
 
         tx.commit()?;
@@ -331,6 +340,14 @@ impl Index {
 
   pub(crate) fn has_sat_index(&self) -> Result<bool> {
     match self.begin_read()?.0.open_table(OUTPOINT_TO_SAT_RANGES) {
+      Ok(_) => Ok(true),
+      Err(redb::TableError::TableDoesNotExist(_)) => Ok(false),
+      Err(err) => Err(err.into()),
+    }
+  }
+
+  pub(crate) fn has_utxo_index(&self) -> Result<bool> {
+    match self.begin_read()?.0.open_table(SAT_TO_OUTPOINT) {
       Ok(_) => Ok(true),
       Err(redb::TableError::TableDoesNotExist(_)) => Ok(false),
       Err(err) => Err(err.into()),
@@ -717,6 +734,13 @@ impl Index {
     }
   }
 
+  pub(crate) fn get_transaction_info(
+    &self,
+    txid: Txid,
+  ) -> Result<GetRawTransactionResult, bitcoincore_rpc::Error> {
+    self.client.get_raw_transaction_info(&txid, None)
+  }
+
   pub(crate) fn get_transaction_blockhash(&self, txid: Txid) -> Result<Option<BlockHash>> {
     Ok(
       self
@@ -744,65 +768,16 @@ impl Index {
     )
   }
 
-  pub(crate) fn find(&self, sat: u64, outpoints: &Vec<OutPoint>) -> Result<Option<SatPoint>> {
-    self.require_sat_index("find")?;
-
-    let rtx = self.begin_read()?;
-
-    if rtx.block_count()? <= Sat(sat).height().n() {
-      return Ok(None);
-    }
-
-    if outpoints.is_empty() {
-      let outpoint_to_sat_ranges = rtx.0.open_table(OUTPOINT_TO_SAT_RANGES)?;
-
-      for range in outpoint_to_sat_ranges.range::<&[u8; 36]>(&[0; 36]..)? {
-        let (key, value) = range?;
-        let mut offset = 0;
-        for chunk in value.value().chunks_exact(11) {
-          let (start, end) = SatRange::load(chunk.try_into().unwrap());
-          if start <= sat && sat < end {
-            return Ok(Some(SatPoint {
-              outpoint: Entry::load(*key.value()),
-              offset: offset + sat - start,
-            }));
-          }
-          offset += end - start;
-        }
-      }
-
-      Ok(None)
-    } else {
-      for outpoint in outpoints {
-        match self.list(*outpoint)? {
-          Some(crate::index::List::Unspent(ranges)) => {
-            let mut offset = 0;
-            for (start, end) in ranges {
-              if start <= sat && sat < end {
-                return Ok(Some(SatPoint {
-                  outpoint: *outpoint,
-                  offset: offset + sat - start,
-                }));
-              }
-              offset += end - start;
-            }
-            Ok::<(), Error>(())
-          }
-          Some(crate::index::List::Spent) => Err(anyhow!("output spent.")),
-          None => Err(anyhow!("output not found")),
-        }?;
-      }
-      Ok(None)
-    }
-  }
-
-  pub(crate) fn find_range(
+  pub(crate) fn find(
     &self,
-    search_start: u64,
-    search_end: u64,
+    search_start: Sat,
+    search_end: Sat,
     outpoints: &Vec<OutPoint>,
   ) -> Result<Option<Vec<FindRangeOutput>>> {
     self.require_sat_index("find")?;
+
+    let search_start = search_start.0;
+    let search_end = search_end.0;
 
     let rtx = self.begin_read()?;
 
@@ -812,35 +787,125 @@ impl Index {
 
     let mut result = Vec::new();
     if outpoints.is_empty() {
+      // no outpoints provided; search the whole space
       let mut remaining_sats = search_end - search_start;
       let outpoint_to_sat_ranges = rtx.0.open_table(OUTPOINT_TO_SAT_RANGES)?;
 
-      for range in outpoint_to_sat_ranges.range::<&[u8; 36]>(&[0; 36]..)? {
-        let (key, value) = range?;
-        let mut offset = 0;
-        for chunk in value.value().chunks_exact(11) {
-          let (start, end) = SatRange::load(chunk.try_into().unwrap());
-          if start < search_end && search_start < end {
-            let overlap_start = cmp::max(start, search_start);
-            let overlap_end = cmp::min(search_end, end);
-            result.push(FindRangeOutput {
-              start: overlap_start,
-              size: overlap_end - overlap_start,
-              satpoint: SatPoint {
-                outpoint: Entry::load(*key.value()),
-                offset: offset + overlap_start - start,
-              },
-            });
+      if self.has_utxo_index()? {
+        let sat_to_outpoint = rtx.0.open_table(SAT_TO_OUTPOINT)?;
 
-            remaining_sats -= overlap_end - overlap_start;
-            if remaining_sats == 0 {
-              break;
+        // loop through all the sats backwards from the end of the search range
+        'outer: for range in sat_to_outpoint.range::<u64>(0..search_end)?.rev() {
+          let (sat, value) = range?;
+          let sat = sat.value();
+          let value = *value.value();
+          let first_outpoint = OutPointPrefix::load(value);
+          let last_outpoint = outpoint_prefix_end(value);
+          tprintln!("| sat {} value {:02x?}", sat, value);
+
+          // loop through all the outputs that match the 2 bytes we looked up
+          'inner: for range in
+            outpoint_to_sat_ranges.range::<&[u8; 36]>(&first_outpoint..=&last_outpoint)?
+          {
+            let (key, value) = range?;
+            let mut offset = 0;
+            tprintln!(
+              "| | found output: {:?}",
+              <OutPoint as Entry>::load(*key.value())
+            );
+
+            // loop through the ranges in the output
+            for chunk in value.value().chunks_exact(11) {
+              let (start, end) = SatRange::load(chunk.try_into().unwrap());
+
+              tprintln!("| | | found range in output: {} - {}", start, end);
+
+              // our range will start with the key we found in SAT_TO_OUTPOINT
+              if start == sat {
+                tprintln!("| | | this is our range");
+                if start < search_end && search_start < end {
+                  let overlap_start = cmp::max(start, search_start);
+                  let overlap_end = cmp::min(search_end, end);
+                  tprintln!(
+                    "| | | it overlaps from {} to {}",
+                    overlap_start,
+                    overlap_end
+                  );
+                  result.push(FindRangeOutput {
+                    start: overlap_start,
+                    size: overlap_end - overlap_start,
+                    satpoint: SatPoint {
+                      outpoint: Entry::load(*key.value()),
+                      offset: offset + overlap_start - start,
+                    },
+                  });
+
+                  remaining_sats -= overlap_end - overlap_start;
+                  tprintln!(
+                    "| | | remaining down by {} to {}",
+                    overlap_end - overlap_start,
+                    remaining_sats
+                  );
+
+                  if remaining_sats == 0 {
+                    tprintln!("| | | should break 'outer;");
+                    break 'outer;
+                  }
+                } else {
+                  panic!("no overlap");
+                }
+
+                // we found our range. we don't need to look at the other outputs with the same 2 byte start
+                break 'inner;
+              } else {
+                tprintln!("| | | not our range");
+              }
+
+              tprintln!(
+                "| | | offset up {} from {} to {}",
+                end - start,
+                offset,
+                offset + end - start
+              );
+              offset += end - start;
+            } // loop 3 - for each range in the output
+            tprintln!("| | that's it for this output - try the next output");
+          } // loop 2 - for each output that fits the pattern
+          tprintln!("| that's all the outputs - try the next sat");
+        } // loop 1 - for each 2 sat and 2 byte pattern
+        tprintln!("that's all the sats");
+        result.reverse();
+      } else {
+        for range in outpoint_to_sat_ranges.range::<&[u8; 36]>(&[0; 36]..)? {
+          let (key, value) = range?;
+          let mut offset = 0;
+          for chunk in value.value().chunks_exact(11) {
+            let (start, end) = SatRange::load(chunk.try_into().unwrap());
+            if start < search_end && search_start < end {
+              let overlap_start = cmp::max(start, search_start);
+              let overlap_end = cmp::min(search_end, end);
+              result.push(FindRangeOutput {
+                start: overlap_start,
+                size: overlap_end - overlap_start,
+                satpoint: SatPoint {
+                  outpoint: Entry::load(*key.value()),
+                  offset: offset + overlap_start - start,
+                },
+              });
+
+              remaining_sats -= overlap_end - overlap_start;
+              if remaining_sats == 0 {
+                break;
+              }
             }
+            offset += end - start;
           }
-          offset += end - start;
         }
       }
     } else {
+      // outpoints provided; only search the outpoints
+      let mut result_map = BTreeMap::new();
+
       for outpoint in outpoints {
         match self.list(*outpoint)? {
           Some(crate::index::List::Unspent(ranges)) => {
@@ -849,14 +914,17 @@ impl Index {
               if start < search_end && search_start < end {
                 let overlap_start = cmp::max(start, search_start);
                 let overlap_end = cmp::min(search_end, end);
-                result.push(FindRangeOutput {
-                  start: overlap_start,
-                  size: overlap_end - overlap_start,
-                  satpoint: SatPoint {
-                    outpoint: *outpoint,
-                    offset: offset + overlap_start - start,
+                result_map.insert(
+                  overlap_start,
+                  FindRangeOutput {
+                    start: overlap_start,
+                    size: overlap_end - overlap_start,
+                    satpoint: SatPoint {
+                      outpoint: *outpoint,
+                      offset: offset + overlap_start - start,
+                    },
                   },
-                });
+                );
               }
               offset += end - start;
             }
@@ -866,6 +934,8 @@ impl Index {
           None => Err(anyhow!("output not found")),
         }?;
       }
+
+      result = result_map.into_values().collect();
     }
 
     Ok(Some(result))
@@ -1124,6 +1194,10 @@ impl Index {
         .get(&inscription_id.store())?
         .map(|value| InscriptionEntry::load(value.value())),
     )
+  }
+
+  pub(crate) fn compact_db(&mut self) -> Result<bool, CompactionError> {
+    self.database.compact()
   }
 
   pub(crate) fn delete_transfer_log(&self) -> Result {
@@ -1746,13 +1820,21 @@ mod tests {
   fn find_first_sat() {
     let context = Context::builder().arg("--index-sats").build();
     assert_eq!(
-      context.index.find(0, &Vec::new()).unwrap().unwrap(),
-      SatPoint {
-        outpoint: "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b:0"
-          .parse()
-          .unwrap(),
-        offset: 0,
-      }
+      context
+        .index
+        .find(Sat(0), Sat(1), &Vec::new())
+        .unwrap()
+        .unwrap(),
+      vec![FindRangeOutput {
+        start: 0,
+        size: 1,
+        satpoint: SatPoint {
+          outpoint: "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b:0"
+            .parse()
+            .unwrap(),
+          offset: 0,
+        }
+      }]
     )
   }
 
@@ -1760,13 +1842,21 @@ mod tests {
   fn find_second_sat() {
     let context = Context::builder().arg("--index-sats").build();
     assert_eq!(
-      context.index.find(1, &Vec::new()).unwrap().unwrap(),
-      SatPoint {
-        outpoint: "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b:0"
-          .parse()
-          .unwrap(),
-        offset: 1,
-      }
+      context
+        .index
+        .find(Sat(1), Sat(2), &Vec::new())
+        .unwrap()
+        .unwrap(),
+      vec![FindRangeOutput {
+        start: 1,
+        size: 1,
+        satpoint: SatPoint {
+          outpoint: "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b:0"
+            .parse()
+            .unwrap(),
+          offset: 1,
+        }
+      }]
     )
   }
 
@@ -1777,15 +1867,19 @@ mod tests {
     assert_eq!(
       context
         .index
-        .find(50 * COIN_VALUE, &Vec::new())
+        .find(Sat(50 * COIN_VALUE), Sat(50 * COIN_VALUE + 1), &Vec::new())
         .unwrap()
         .unwrap(),
-      SatPoint {
-        outpoint: "30f2f037629c6a21c1f40ed39b9bd6278df39762d68d07f49582b23bcb23386a:0"
-          .parse()
-          .unwrap(),
-        offset: 0,
-      }
+      vec![FindRangeOutput {
+        start: 50 * COIN_VALUE,
+        size: 1,
+        satpoint: SatPoint {
+          outpoint: "30f2f037629c6a21c1f40ed39b9bd6278df39762d68d07f49582b23bcb23386a:0"
+            .parse()
+            .unwrap(),
+          offset: 0,
+        }
+      }]
     )
   }
 
@@ -1793,7 +1887,10 @@ mod tests {
   fn find_unmined_sat() {
     let context = Context::builder().arg("--index-sats").build();
     assert_eq!(
-      context.index.find(50 * COIN_VALUE, &Vec::new()).unwrap(),
+      context
+        .index
+        .find(Sat(50 * COIN_VALUE), Sat(50 * COIN_VALUE + 1), &Vec::new())
+        .unwrap(),
       None
     );
   }
@@ -1811,13 +1908,17 @@ mod tests {
     assert_eq!(
       context
         .index
-        .find(50 * COIN_VALUE, &Vec::new())
+        .find(Sat(50 * COIN_VALUE), Sat(50 * COIN_VALUE + 1), &Vec::new())
         .unwrap()
         .unwrap(),
-      SatPoint {
-        outpoint: OutPoint::new(spend_txid, 0),
-        offset: 0,
-      }
+      vec![FindRangeOutput {
+        start: 50 * COIN_VALUE,
+        size: 1,
+        satpoint: SatPoint {
+          outpoint: OutPoint::new(spend_txid, 0),
+          offset: 0,
+        }
+      }]
     )
   }