+++ src/subcommand/find.rs
@@ -1,18 +1,55 @@
-use super::*;
+use {
+  super::*,
+  chrono::NaiveDateTime,
+  std::io::{BufRead, BufReader},
+};
 
 #[derive(Debug, Parser)]
 pub(crate) struct Find {
   #[clap(long, help = "Only look in specified outpoint(s).")]
   outpoint: Vec<OutPoint>,
+  #[clap(
+    long,
+    help = "Read a list of sats and ranges to find from a file. One sat or range per line. Ranges written as <start>-<end>."
+  )]
+  file: Vec<PathBuf>,
+  #[clap(long, help = "Ignore bad sat ranges.")]
+  ignore: bool,
+  #[clap(long, help = "Show addresses in the results.")]
+  show_address: bool,
+  #[clap(long, help = "Show blockhashes in the results.")]
+  show_blockhash: bool,
+  #[clap(long, help = "Show dates in the results.")]
+  show_date: bool,
+  #[clap(long, help = "Show sat names in the results.")]
+  show_name: bool,
+  #[clap(long, help = "Show timestamps in the results.")]
+  show_time: bool,
+  #[clap(long, help = "Show output values in the results.")]
+  show_value: bool,
   #[clap(help = "Find output and offset of <SAT>.")]
-  sat: Sat,
+  sat: Option<Sat>,
   #[clap(help = "Find output and offset of all sats in the range <SAT>-<END>.")]
   end: Option<Sat>,
 }
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Output {
+  pub start: u64,
+  pub size: u64,
   pub satpoint: SatPoint,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub address: Option<Address<NetworkUnchecked>>,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub blockhash: Option<bitcoin::BlockHash>,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub date: Option<String>,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub name: Option<String>,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub timestamp: Option<usize>,
+  #[serde(default, skip_serializing_if = "Option::is_none")]
+  pub value: Option<u64>,
 }
 
 impl Find {
@@ -21,32 +58,174 @@ impl Find {
 
     index.update()?;
 
-    match self.end {
-      Some(end) => {
-        if self.sat < end {
-          match index.find_range(self.sat.0, end.0, &self.outpoint)? {
-            Some(result) => {
-              print_json(result)?;
-              Ok(())
+    let mut targets = Vec::new();
+    let mut results = Vec::new();
+
+    if let Some(sat) = self.sat {
+      let end = match self.end {
+        Some(end) => end,
+        None => sat + 1,
+      };
+
+      if sat < end {
+        targets.push((sat, end));
+      } else {
+        bail!("range is empty");
+      }
+    }
+
+    let comment_re = Regex::new(r"#.*")?;
+    for file in self.file {
+      let reader = BufReader::new(File::open(&file)?);
+      let mut line_number = 0;
+      for line in reader.lines() {
+        line_number += 1;
+        let line = line?;
+        let line = comment_re.replace(&line, "");
+        let line = line.trim();
+        if line.is_empty() {
+          continue;
+        }
+        let mut split = line
+          .trim_start_matches('\u{feff}')
+          .split(&['-', '\u{2013}']); // ASCII hyphen or Unicode 'EN DASH'
+        if let Some(start) = split.next() {
+          let start = match start.parse() {
+            Ok(start) => start,
+            Err(err) => bail!(
+              "parsing sat '{start}' on line {line_number} of '{:?}': {err}",
+              file
+            ),
+          };
+
+          let end = match split.next() {
+            Some(end) => {
+              if let Some(junk) = split.next() {
+                bail!(
+                  "trailing junk on line {line_number} of '{}' ({})",
+                  file.display(),
+                  junk
+                );
+              } else {
+                match end.parse() {
+                  Ok(end) => end,
+                  Err(err) => bail!(
+                    "parsing range end '{end}' on line {line_number} of '{:?}': {err}",
+                    file
+                  ),
+                }
+              }
             }
-            None => Err(anyhow!("range has not been mined as of index height")),
-          }
+            None => start + 1,
+          };
+          targets.push((start, end));
         } else {
-          Err(anyhow!("range is empty"))
+          bail!(
+            "file '{}' is not formatted correctly at line {line_number}",
+            file.display()
+          );
+        }
+      }
+    }
+
+    if targets.is_empty() {
+      bail!("nothing to find");
+    }
+
+    // loop through targets
+    for (sat, end) in targets {
+      // eprintln!("find {sat}-{end}");
+      match index.find(sat, end, &self.outpoint)? {
+        Some(result) => {
+          // eprintln!("  found {} satpoints", result.len());
+          results.extend(result);
+        }
+        None => {
+          if !self.ignore {
+            return Err(anyhow!(
+              "range {sat}-{end} not found; use --ignore to continue anyway"
+            ));
+          }
+        }
+      }
+    }
+
+    let mut detailed_results = Vec::new();
+
+    // let gbt = options.chain().genesis_block().coinbase().unwrap().clone();
+    // print_json(&gbt)?;
+    // println!("gbt.output = {:?}", options.chain().address_from_script(&gbt.output[0].script_pubkey));
+    // result.satpoint.outpoint.txid == gbt.txid()
+
+    for result in results {
+      let tx = if self.show_address
+        || self.show_blockhash
+        || self.show_date
+        || self.show_time
+        || self.show_value
+      {
+        index
+          .get_transaction_info(result.satpoint.outpoint.txid)
+          .ok()
+      } else {
+        None
+      };
+
+      let mut result = Output {
+        start: result.start,
+        size: result.size,
+        satpoint: result.satpoint,
+        address: None,
+        blockhash: None,
+        date: None,
+        name: None,
+        timestamp: None,
+        value: None,
+      };
+
+      if let Some(tx) = tx.clone() {
+        if self.show_address {
+          result.address = tx.vout[result.satpoint.outpoint.vout as usize]
+            .script_pub_key
+            .address
+            .clone();
+        }
+
+        if self.show_blockhash {
+          result.blockhash = tx.blockhash;
+        }
+
+        if self.show_date {
+          result.date = Some(
+            NaiveDateTime::from_timestamp_opt(tx.time.unwrap() as i64, 0)
+              .unwrap()
+              .format("%Y-%m-%d %H:%M:%S")
+              .to_string(),
+          );
+        }
+
+        if self.show_time {
+          result.timestamp = tx.time;
+        }
+
+        if self.show_value {
+          result.value = Some(
+            tx.vout[result.satpoint.outpoint.vout as usize]
+              .value
+              .to_sat(),
+          );
         }
       }
 
-      None => match index.find(self.sat.0, &self.outpoint)? {
-        Some(satpoint) => {
-          print_json(Output { satpoint })?;
-          Ok(())
-        }
-        None => Err(anyhow!(if self.outpoint.is_empty() {
-          "sat has not been mined as of index height"
-        } else {
-          "sat was not found in satpoint(s)"
-        })),
-      },
+      if self.show_name {
+        result.name = Some(Sat(result.start).name());
+      }
+
+      detailed_results.push(result);
     }
+
+    print_json(detailed_results)?;
+
+    Ok(())
   }
 }