+++ src/subcommand/wallet/inscribe.rs
@@ -402,9 +402,23 @@ impl Inscribe {
 
         if !signed_reveal_tx.complete {
           return Err(anyhow!(
-            "error signing commit tx: {:?}",
+            "error signing reveal tx: {:?}",
             signed_reveal_tx.errors
           ));
+        }
+
+        let reveal_weight = client
+          .call::<DecodeRawTransactionOutput>(
+            "decoderawtransaction",
+            &[signed_reveal_tx.hex.raw_hex().into()],
+          )?
+          .weight;
+
+        if !self.no_limit && reveal_weight > bitcoin::Weight::from_wu(MAX_STANDARD_TX_WEIGHT.into())
+        {
+          bail!(
+            "reveal transaction weight greater than {MAX_STANDARD_TX_WEIGHT} (MAX_STANDARD_TX_WEIGHT): {reveal_weight}"
+          );
         }
 
         signed_reveal_txs.push((reveal_tx, signed_reveal_tx.hex));
@@ -630,7 +644,7 @@ impl Inscribe {
         .keys()
         .find(|outpoint| {
           !inscribed_utxos.contains(outpoint)
-            && (cursed_outpoint.is_none() || **outpoint != cursed_outpoint.clone().unwrap())
+            && (cursed_outpoint.is_none() || **outpoint != cursed_outpoint.unwrap())
         })
         .map(|outpoint| SatPoint {
           outpoint: *outpoint,
@@ -707,9 +721,9 @@ impl Inscribe {
         value: 0,
       }];
 
-      if cursed_outpoint.is_some() {
+      if let Some(cursed_outpoint) = cursed_outpoint {
         let cursed_txout = cursed_txout.as_ref().unwrap();
-        inputs.insert(0, cursed_outpoint.unwrap());
+        inputs.insert(0, cursed_outpoint);
         outputs.insert(
           0,
           TxOut {
@@ -733,8 +747,8 @@ impl Inscribe {
     }
 
     let mut utxos_clone = utxos.clone();
-    if cursed_outpoint.is_some() {
-      utxos_clone.remove(&cursed_outpoint.unwrap());
+    if let Some(cursed_outpoint) = cursed_outpoint {
+      utxos_clone.remove(&cursed_outpoint);
     }
 
     tprintln!("[make commit]");
@@ -776,9 +790,9 @@ impl Inscribe {
         value: output.value,
       }];
 
-      if cursed_outpoint.is_some() {
+      if let Some(cursed_outpoint) = cursed_outpoint {
         let cursed_txout = cursed_txout.as_ref().unwrap();
-        inputs.insert(0, cursed_outpoint.unwrap());
+        inputs.insert(0, cursed_outpoint);
         outputs.insert(
           0,
           TxOut {
@@ -835,7 +849,7 @@ impl Inscribe {
       let signature = secp256k1.sign_schnorr(
         &secp256k1::Message::from_slice(signature_hash.as_ref())
           .expect("should be cryptographically secure hash"),
-        &key_pair,
+        key_pair,
       );
 
       let witness = sighash_cache
@@ -992,6 +1006,8 @@ mod tests {
         [commit_address, change(1)],
         reveal_address,
         None,
+        None,
+        None,
         FeeRate::try_from(1.0).unwrap(),
         FeeRate::try_from(1.0).unwrap(),
         None,
@@ -1028,6 +1044,8 @@ mod tests {
       utxos.into_iter().collect(),
       [commit_address, change(1)],
       reveal_address,
+      None,
+      None,
       None,
       FeeRate::try_from(1.0).unwrap(),
       FeeRate::try_from(1.0).unwrap(),
@@ -1069,6 +1087,8 @@ mod tests {
       utxos.into_iter().collect(),
       [commit_address, change(1)],
       reveal_address,
+      None,
+      None,
       None,
       FeeRate::try_from(1.0).unwrap(),
       FeeRate::try_from(1.0).unwrap(),
@@ -1118,6 +1138,8 @@ mod tests {
       [commit_address, change(1)],
       reveal_address,
       None,
+      None,
+      None,
       FeeRate::try_from(1.0).unwrap(),
       FeeRate::try_from(1.0).unwrap(),
       None,
@@ -1160,6 +1182,8 @@ mod tests {
         utxos.into_iter().collect(),
         [commit_address, change(1)],
         reveal_address,
+        None,
+        None,
         None,
         FeeRate::try_from(fee_rate).unwrap(),
         FeeRate::try_from(fee_rate).unwrap(),
@@ -1232,6 +1256,8 @@ mod tests {
         [commit_address, change(1)],
         reveal_address,
         None,
+        None,
+        None,
         FeeRate::try_from(commit_fee_rate).unwrap(),
         FeeRate::try_from(fee_rate).unwrap(),
         None,
@@ -1287,6 +1313,8 @@ mod tests {
       [commit_address, change(1)],
       reveal_address,
       None,
+      None,
+      None,
       FeeRate::try_from(1.0).unwrap(),
       FeeRate::try_from(1.0).unwrap(),
       None,
@@ -1324,6 +1352,8 @@ mod tests {
         utxos.into_iter().collect(),
         [commit_address, change(1)],
         reveal_address,
+        None,
+        None,
         None,
         FeeRate::try_from(1.0).unwrap(),
         FeeRate::try_from(1.0).unwrap(),